Move circular rebalance and swap inhibit logic
