+++ Boss/Mod/FundsMover/Main.cpp
@@ -6,7 +6,10 @@
 #include"Boss/ModG/RebalanceUnmanagerProxy.hpp"
 #include"Boss/Msg/Init.hpp"
 #include"Boss/Msg/ProvideDeletablePaymentLabelFilter.hpp"
+#include"Boss/Msg/RequestGetCircularRebalanceFlag.hpp"
 #include"Boss/Msg/RequestMoveFunds.hpp"
+#include"Boss/Msg/ResponseGetCircularRebalanceFlag.hpp"
+#include"Boss/Msg/ResponseMoveFunds.hpp"
 #include"Boss/Msg/SolicitDeletablePaymentLabelFilter.hpp"
 #include"Boss/concurrent.hpp"
 #include"Boss/log.hpp"
@@ -29,6 +32,11 @@ private:
 	Claimer claimer;
 	Boss::Mod::Rpc* rpc;
 	Ln::NodeId self_id;
+	Msg::ResponseGetCircularRebalanceFlag resp_circular_rebalance_flag;
+
+	ModG::ReqResp< Msg::RequestGetCircularRebalanceFlag
+		     , Msg::ResponseGetCircularRebalanceFlag
+		     > get_circular_rebalance_rr;
 
 	Boss::ModG::RebalanceUnmanagerProxy unmanager;
 
@@ -40,47 +48,62 @@ private:
 		});
 		bus.subscribe<Msg::RequestMoveFunds
 			     >([this](Msg::RequestMoveFunds const& m) {
-			auto msg = std::make_shared<Msg::RequestMoveFunds>(m);
-			return wait_for_rpc().then([this]() {
-				return unmanager.get_unmanaged();
-			}).then([this, msg](std::set<Ln::NodeId> const* unmanaged) {
-				auto un_s = (unmanaged->count(msg->source) != 0);
-				auto un_d = (unmanaged->count(msg->destination) != 0);
-				if (un_s || un_d) {
-					char const* tpl = nullptr;
-					if (un_s && un_d) {
-						tpl = "%1$sfrom an unmanaged node %2$s "
-						      "to an unmanaged node %3$s%4$s"
-						      ;
-					} else if (un_s) {
-						tpl = "%1$sfrom an unmanaged node %2$s"
-						      "%4$s"
-						      ;
-					} else {
-						tpl = "%1$s"
-						      "to an unmanaged node %3$s%4$s"
-						      ;
+			return get_circular_rebalance_rr.execute(
+					Msg::RequestGetCircularRebalanceFlag{
+				nullptr
+			}).then([this, m](Msg::ResponseGetCircularRebalanceFlag res) {
+				resp_circular_rebalance_flag = res;
+				auto msg = std::make_shared<Msg::RequestMoveFunds>(m);
+				return wait_for_rpc().then([this]() {
+					return unmanager.get_unmanaged();
+				}).then([this, m, msg](std::set<Ln::NodeId> const* unmanaged) {
+					auto un_s = (unmanaged->count(msg->source) != 0);
+					auto un_d = (unmanaged->count(msg->destination) != 0);
+					if (un_s || un_d) {
+						char const* tpl = nullptr;
+						if (un_s && un_d) {
+							tpl = "%1$sfrom an unmanaged node %2$s "
+							      "to an unmanaged node %3$s%4$s"
+							      ;
+						} else if (un_s) {
+							tpl = "%1$sfrom an unmanaged node %2$s"
+							      "%4$s"
+							      ;
+						} else {
+							tpl = "%1$s"
+							      "to an unmanaged node %3$s%4$s"
+							      ;
+						}
+						return Boss::log( bus, Error
+								, tpl
+								, "FundsMover: *SOMETHING* is "
+								  "attempting to move funds "
+								, Util::stringify(msg->source)
+									.c_str()
+								, Util::stringify(msg->destination)
+									.c_str()
+								, ", this may be a bug, "
+								  "refusing to move.  "
+								  "Contact " PACKAGE_BUGREPORT
+								);
 					}
-					return Boss::log( bus, Error
-							, tpl
-							, "FundsMover: *SOMETHING* is "
-							  "attempting to move funds "
-							, Util::stringify(msg->source)
-								.c_str()
-							, Util::stringify(msg->destination)
-								.c_str()
-							, ", this may be a bug, "
-							  "refusing to move.  "
-							  "Contact " PACKAGE_BUGREPORT
-							);
-				}
-				auto runner = Runner::create( bus
-							    , *rpc
-							    , self_id
-							    , claimer
-							    , *msg
-							    );
-				return Runner::start(runner);
+					if (!resp_circular_rebalance_flag.state)
+						return Boss::log( bus, Info
+								, "FundsMover: %s."
+								, resp_circular_rebalance_flag.comment.c_str()
+								)
+							+ bus.raise(Msg::ResponseMoveFunds{
+								m.requester, Ln::Amount(), Ln::Amount()
+							  })
+							;
+					auto runner = Runner::create( bus
+								    , *rpc
+								    , self_id
+								    , claimer
+								    , *msg
+								    );
+					return Runner::start(runner);
+				});
 			});
 		});
 		using Msg::ProvideDeletablePaymentLabelFilter;
@@ -110,6 +133,7 @@ public:
 	Impl(S::Bus& bus_) : bus(bus_)
 			   , claimer(bus_)
 			   , rpc(nullptr)
+			   , get_circular_rebalance_rr(bus_)
 			   , unmanager(bus_)
 			   { start(); }
 };
